# ResearchEngine.Blazor/openapi/v3.yaml
openapi: 3.1.1
info:
  title: ResearchEngine.Web | v1
  version: 1.0.0
servers:
  - url: http://localhost:8090/
paths:
...
  /api/protocol/clarifications:
    post:
      tags:
        - Research Protocol API
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProtocolClarificationsRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProtocolClarificationsResponse"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /api/protocol/parameters:
    post:
      tags:
        - Research Protocol API
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProtocolParametersRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProtocolParametersResponse"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
components:
  schemas:
...
    ClarificationDto:
      required:
        - question
        - answer
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
...
    ProblemDetails:
      type: object
      properties:
        type:
          type:
            - "null"
            - string
        title:
          type:
            - "null"
            - string
        status:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          format: int32
        detail:
          type:
            - "null"
            - string
        instance:
          type:
            - "null"
            - string
    ProtocolClarificationsRequest:
      required:
        - query
      type: object
      properties:
        query:
          type: string
        includeConfigureQuestions:
          type: boolean
          default: false
    ProtocolClarificationsResponse:
      required:
        - questions
      type: object
      properties:
        questions:
          type: array
          items:
            type: string
    ProtocolParametersRequest:
      required:
        - query
        - clarifications
        - overrides
      type: object
      properties:
        query:
          type: string
        clarifications:
          type:
            - "null"
            - array
          items:
            $ref: "#/components/schemas/ClarificationDto"
        overrides:
          type:
            - "null"
            - object
    ProtocolParametersResponse:
      required:
        - breadth
        - depth
        - language
        - region
      type: object
      properties:
        breadth:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        depth:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        language:
          type: string
        region:
          type:
            - "null"
            - string
...